Added Firefox Version
The extension itself has not been modified. A version for Firefox 29
onwards has been created and added to this repository. This version uses
the Firefox add on SDK (latest development version and not the release
version (currently 1.15)

At the time of this commit. The extension only works with Firefox 29
onwards (currently beta). To run the extension in development mode, you
need the latest development version and not the release version. See
https://developer.mozilla.org/en-US/Add-ons/SDK/Tutorials/Installation#Add-on_SDK_Git_Repository